# docs/swagger.yaml
basePath: /
definitions:
  main.driverLoginRequest:
    properties:
      password:
        type: string
      phone:
        type: string
    required:
    - password
    - phone
    type: object
host: localhost:9080
info:
  contact: {}
  description: |-
    BFF (Backend for Frontend) that orchestrates user-service, driver-service, and location-service.
    Handles auth, driver profile merging, and proxies to downstream services.
  title: StudEx API Gateway
  version: "1.0"
paths:
  /drivers/auth:
    post:
      consumes:
      - application/json
      description: Authenticate driver via user-service, fetch driver profile, merge
        user data
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.driverLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Driver login
      tags:
      - auth
  /drivers/me:
    get:
      description: Get driver profile enriched with user data (name, display_name,
        gender)
      parameters:
      - description: Driver phone number
        in: query
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get driver profile
      tags:
      - driver
schemes:
- http
- https
swagger: "2.0"